Move Rust resource primitives into core
