+++ .ralph/rust-bindings-plan.state.json
@@ -1,7 +1,7 @@
 {
   "name": "rust-bindings-plan",
   "taskFile": ".ralph/rust-bindings-plan.md",
-  "iteration": 21,
+  "iteration": 24,
   "maxIterations": 50,
   "itemsPerIteration": 3,
   "reflectEvery": 10,
